+++ facetime/src/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./PermissionModal";
+export * from './DarkModeToggle';
