+++ facetime/src/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from "./useFetchToken";
+export * from './useDarkMode';
